RF: consolidate dep checking and inheritance in `siblings`
